แก้redirect ที่ผิด, แก้bugลบร้านค้า

+++ protected/modules/webSimulation/controllers/ManageShopController.php
@@ -21,7 +21,6 @@ class ManageShopController extends Controller {
     //หน้า register
     public function actionRegister() {
         Yii::app()->clientScript->registerScriptFile(Yii::app()->request->baseUrl . '/js/self/web_sim/shop_register.js');
-        $shop_id == NULL;
         $shop_id = Yii::app()->session['shop_id'];
         if ($shop_id == NULL) {
             $model = new WebShop();
@@ -64,6 +63,7 @@ class ManageShopController extends Controller {
         $shop_format_id = WebShopFormat::model()->findByAttributes(array('web_shop_id' => $shop_id))->web_shop_format_id;
         if ($shop_format_id == NULL) {
             $model = new WebShopFormat();
+            $is_new = TRUE;
         } else {
             $model = WebShopFormat::model()->findByPk($shop_format_id);
         }
@@ -73,7 +73,7 @@ class ManageShopController extends Controller {
 
             if ($model->theme != '') {
                 $model->save();
-                if ($shop_id == NULL) {
+                if ($is_new) {
                     //หน้าสมัคร
                     echo "<script language='javascript'>
                         alert('" . Yii::t('language', 'บันทึก') . Yii::t('language', 'ข้อมูล') . Yii::t('language', 'เรียบร้อย') . "');
@@ -116,8 +116,12 @@ class ManageShopController extends Controller {
         $model = WebShop::model()->findByPk($shop_id);
         $format = WebShopFormat::model()->findByAttributes(array('web_shop_id' => $shop_id));
         if ($model->delete()) {
-            unlink(Yii::app()->basePath . $format->logo);
-            unlink(Yii::app()->basePath . $format->background);
+            if ($format->logo != NULL) {
+                unlink(Yii::app()->basePath . $format->logo);
+            }
+            if ($format->background != NULL) {
+                unlink(Yii::app()->basePath . $format->background);
+            }
         }
     }
 
@@ -501,9 +505,31 @@ class ManageShopController extends Controller {
         }
         $this->render('sort_box', array('shop_id' => $shop_id));
     }
-    
-    public function actionEditBox(){
-        
+
+    public function actionEditBox($box_id) {
+        $shop_id = Yii::app()->session['shop_id'];
+//        Yii::app()->clientScript->registerCssFile('http://code.jquery.com/ui/1.10.3/themes/smoothness/jquery-ui.css');
+//        Yii::app()->clientScript->registerScriptFile('http://code.jquery.com/jquery-1.9.1.js');
+//        Yii::app()->clientScript->registerScriptFile('http://code.jquery.com/ui/1.10.3/jquery-ui.js');
+        Yii::app()->clientScript->registerScriptFile(Yii::app()->request->baseUrl . '/js/self/web_sim/edit_item_in_box.js');
+        if ($_POST['select'] != '') {
+//            $criteria = new CDbCriteria;
+//            $criteria->select = 'order_n';
+//            $criteria->limit = '1';
+//            if ($order = $model->model()->find($criteria)) {
+            WebShopBoxItem::model()->deleteAll('web_shop_box_id = ' . $box_id);
+            $arr = array();
+            $arr = preg_split('/,/', $_POST['select']);
+            foreach ($arr as $item_id) {
+                $model = new WebShopBoxItem();
+                $model->web_shop_id = $shop_id;
+                $model->web_shop_box_id = $box_id;
+                $model->web_shop_item_id = $item_id;
+                $model->save();
+            }
+        }
+        $this->render('edit_box', array('box_id' => $box_id, 'shop_id' => $shop_id));
     }
 
 }
+
